Update from upstream repo Kwoth/NadekoBot@1.9

+++ NadekoBot.Core/Modules/Administration/UserPunishCommands.cs
@@ -533,7 +533,7 @@ namespace NadekoBot.Modules.Administration
                             Type = BlacklistType.User,
                         }));
                     //clear their currencies
-                    uow.DiscordUsers.RemoveFromMany(found.Select(x => (long)x).ToList());
+                    uow.DiscordUsers.RemoveFromMany(found.Select(x => x).ToList());
                     uow.Complete();
                 }
