+++ NadekoBot.Core/Services/Database/Repositories/Impl/GuildConfigRepository.cs
@@ -24,9 +24,9 @@ namespace NadekoBot.Core.Services.Database.Repositories.Impl
                 }
             };
 
-        public IEnumerable<GuildConfig> GetAllGuildConfigs(List<long> availableGuilds) =>
+        public IEnumerable<GuildConfig> GetAllGuildConfigs(List<ulong> availableGuilds) =>
             IncludeEverything()
-                .Where(gc => availableGuilds.Contains((long)gc.GuildId))
+                .Where(gc => availableGuilds.Contains(gc.GuildId))
                 .ToList();
 
         private IQueryable<GuildConfig> IncludeEverything()
@@ -130,10 +130,10 @@ namespace NadekoBot.Core.Services.Database.Repositories.Impl
             return config;
         }
 
-        public IEnumerable<GuildConfig> Permissionsv2ForAll(List<long> include)
+        public IEnumerable<GuildConfig> Permissionsv2ForAll(List<ulong> include)
         {
             var query = _set
-                .Where(x => include.Contains((long)x.GuildId))
+                .Where(x => include.Contains(x.GuildId))
                 .Include(gc => gc.Permissions);
 
             return query.ToList();
@@ -172,10 +172,10 @@ namespace NadekoBot.Core.Services.Database.Repositories.Impl
                 .ToArray();
         }
 
-        public IEnumerable<FollowedStream> GetFollowedStreams(List<long> included)
+        public IEnumerable<FollowedStream> GetFollowedStreams(List<ulong> included)
         {
             return _set
-                .Where(gc => included.Contains((long)gc.GuildId))
+                .Where(gc => included.Contains(gc.GuildId))
                 .Include(gc => gc.FollowedStreams)
                 .SelectMany(gc => gc.FollowedStreams)
                 .ToList();
