+++ NadekoBot.Core/Services/NadekoBot.cs
@@ -137,7 +137,7 @@ namespace NadekoBot
 
         private void AddServices()
         {
-            var startingGuildIdList = Client.Guilds.Select(x => (long)x.Id).ToList();
+            var startingGuildIdList = Client.Guilds.Select(x => x.Id).ToList();
 
             //this unit of work will be used for initialization of all modules too, to prevent multiple queries from running
             using (var uow = _db.UnitOfWork)
@@ -478,7 +478,7 @@ namespace NadekoBot
         //        if (_loadedPackages.ContainsKey(name))
         //            return false;
 
-        //        var startingGuildIdList = Client.Guilds.Select(x => (long)x.Id).ToList();
+        //        var startingGuildIdList = Client.Guilds.Select(x => x.Id).ToList();
         //        using (var uow = _db.UnitOfWork)
         //        {
         //            AllGuildConfigs = uow.GuildConfigs.GetAllGuildConfigs(startingGuildIdList).ToImmutableArray();
